+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Repositories/PollsRepository_Mock.swift
@@ -13,7 +13,11 @@ final class PollsRepository_Mock: PollsRepository, Spy, @unchecked Sendable {
     @Atomic var suggestPollOption_completion: ((Error?) -> Void)?
     @Atomic var deletePoll_completion: ((Error?) -> Void)?
     
+    var recordedFunctions: [String] = []
     let spyState: SpyState = .init()
+    
+    // Mock for link method
+    var link: ((PollVote, PollVoteListQuery) -> Void)?
 
     override func queryPollVotes(
         query: PollVoteListQuery,
@@ -63,5 +67,9 @@ final class PollsRepository_Mock: PollsRepository, Spy, @unchecked Sendable {
         completion: ((Error?) -> Void)? = nil
     ) {
         deletePoll_completion = completion
+    }
+    
+    override func link(pollVote: PollVote, to query: PollVoteListQuery) {
+        link?(pollVote, query)
     }
 }